+++ src/utils/bugsnag.ts
@@ -2,7 +2,6 @@ import * as React from "react";
 import Bugsnag from "@bugsnag/js";
 import BugsnagPluginReact from "@bugsnag/plugin-react";
 import { APP_VERSION } from "utils/env";
-import { APP_CONFIG } from "config";
 
 if (APP_CONFIG.bugsnagApiKey) {
   Bugsnag.start({